requirements.txt
langchain
faiss-cpu
google-generativeai
pypdf
python-dotenv
bert-score
pandas
numpy
tiktoken
langchain-community
langchain-google-genai
sentence-transformers
scikit-learn
